Add better errors to work around error pages not working
Adding this as a stop gap, we probably want it anyway.

Could be a deeper underlying issues, as accoring to this

http://www.jagregory.com/writings/rails-500-error-showing-in-dev/

In his case, Enumerable#sum had been patched, and it had this far reaching
undesirable result because it was crashing the error page which renders the
errors, thus falling back tot he static error page. I wish I had descovered some
what to detect this, or see what the error is. Anyway, this works around it.

I suspect of yaml_frontmatter, may want to create our own representaion using a
taml header.