New and icon, news and and images js
Just getting started on this, sorta a weird commit. Just cleaning the workspace
and moving on.

The idea is have new be under the root, might not be a great idea as it is
already causing conflicts with the news url, which I'm adding in this commit
too.

Using a flash icon might be cool, might be too much. Flash could be a name for
the product too.