Proof of concept render_layout helper working
This ended up being more tricky that neccessary. I actually arrived at what I
beleieve was a working a superior solution earlier, but but had a oversight that
caused be to beleieve there was was a problem with the order in which things
were rendering. Since this is working as I wanted it (albite a bit crufty) I'm
checking it in, and I will revist to clean up and see if I can arrive at
something more elegant that works equally well.

What this allows for, is to use a partial as a layout that can accept a block,
and render within. This handy for centralizing wrapper code, and creates a
layout within a layout, as it predominatly effects the layout_modules. According
to my budding thoughts on this, this may end up being even more versitile that
for jsut layouts (However not is it curent form as it relies on shared instance
variables.