Introduce yield_to
This is the first take on a stratgey that allows you to specif the layout that
will serve as a container for the content while rendering the content. This also
implies that the layout will be rendered after the template; with this
inversion, vaule can essentially be passed up to the layout from the contained
view, which is usally what you want essentially enabling parameterize layout
that can be constructed to any depth with hope.