Salavage parts from news project, add crawdad, no rivers, look mom, NO RIVERS!
Still buggy for the moment, but a big exciting step nonetheless.

To cherish a moment,
To note a time.
To bring a rhyme,
and to enjoy some wine.

Eat a burger, and go to the store.
Young and old, rich or poor.

Pour pour the rain, rain the knowledge into my brain.
Plant it in the ground, and grow a tree.
Three trees, tress for me to see.

With my eyes, shall I receive a prize?
A prize, a prize, for me to prize.

Wiz, wiz, wiz all about. Bring the song and sing all about.

-- Blake Frost, 2014