Implement full cycle compile with real time update
State machine after_transition hook with mongoid was not working. Issue is
documented here https://github.com/pluginaweek/state_machine/issues/277

Used the version of state_machine from
https://github.com/aganov/state_machine/tree/88e45a1b1622d7748933bc941e7e9b08b3045a94,
referenced in the github issue, to get things working for the time being.

I'm using the transtion to trigger the broadcast of the update to the compiled
version.