Add properties panel for photos, fix js error
There was a porition of code only meant for the when creating a new edition, and
to allow the user to select a publication preset. This was showing on the
editing screens as well, which caused a problem. Modified so that it only shows
on the new screen. Could be cleaner