Update some gems, and experimenting with getting corrdinates of page bounding box
      xCorrection = Math.round(($(window).scrollLeft()/@zoomLevel) * (@zoomLevel - 1))

I use this forumula to correct for error that shows up in the jquery offset
method of the x value. Not sure why, or how browser dependant this is, or even
if it is still a bug in the most recent version. Works for now, altough quite
weird, and creates and creates a leaky abstraction, because, now, to objects are
involved with getting the correct corrdinate values. Also, really poorly written
right now, as these values are all recomputed on each check.