Requesting template from the backend for inserting headlines
This is really hacky. The template should be rendered on the front end, rather
then round tripping and saving the content item up front.

This is nice in order to be able to actually add the headline in realtime and
get the desired effect, but needs to be reworked. Template will need to be front
end compatable, or rather, need to be able to render a template without having
an object saved on the server. That is a resonible server, and will solve the
problem.