Allow drawing of video
Unfortunaly, doesn't quite work because of compatability with the angular
directive responsible for rendering the video.

Need to hook into that some how, because we can't get a reference to the angular
generated elements. Would be ideal to have the clients angular app to initialize
first, this way I can hook into the full rendered document, which would be the
ideal target for the editor. However, there is likley to be many snafus in
supporting these angular directives. Will need to dig further, because this
really is a great way to encapsulate logic and extend the web.