Imply sections/ infront of template name
This is a hacky solution. It should be a little smarter in checking for the
templates existance. Essentially, a template is composed of multiple parts, it
can have a section template part, as well as a masthead section part. This is
really convention over configuration, and used to work around limitations with
content_for