Quick implementation of locking move in the x or y direction
This is a nice start, but I think it needs an idea of direction of move inorder
to do a good job of locking in one direction of the other. When you are close to
the point of the beginning of the drag it is easy to fall off of track, so need
to think more about this. Also currently the way it is coded, we are not showing
snap lines. This should be still enabled.