A real mess of trying to group objects
Getting started on creating groups. Would like to keep the client side, but
needing to go to the server now to get an id for reference. Was trying to create
the guid id locally, but then couldn't save it. This should be improved when we
have better patterns for creating client side, and eventual persistance to the
backend. Further research is required.