Beginning of writing code to ungroup
This is the beginning of ungrouping. Need to be able to properly cleanup views
and make sure to be configured locally.

Will need to do this without touching the backend in the future, to allow for
undo capabilities in the browser, but to get this working, it should be ok.

Also, fixing an issue where group might not have a top or left set, so supply a
default value for now.