Break out CanvasItemView as a module structurally, and mix back into GroupView and ContentItemView, which both inherit it's functionality.
First shot a ruby-like module for mixing funcitnoality. Still working of the
prototype of a coffeescript class even though not using the constructor. I like
class syntax for definition. Would be nice if the was a like version for
`module` which defined a simple class. The tricky part was getting the getters
and setters to copy over, but we are able to do this with the include
implementation which is added to the function prototype, for access during class
definition.