Allow deleting of group, update pages panel
This addition will update the pages panel when leaving edit mode. This is using
the vent to invoke the render. Not sure if this is the stable approach in which
to use. Works for the time being. Ideally, each item in the pages view is
independent, and then it is simple to update the text live, as it is changed.