Create a delete queue, and add groups to that to be deleted as part of a save
This is a nice solution to avoid stale references due to an instantaneous group
delete, and subsequent save of edition that might not happen, leading to dead
group references. By queueing the deletes, we can preform them at the time of
the save, leaving the document in tact on the server until a save occurs, and
helping to ensure that when the save does occur, it will leave the document in a
good state.

A better approach might be doing garbage collection after dereferencing, which
could avoid the need to preform deletes all together, but this works for now,
although not atomic, so may lead to inconsistencies, and doesn't necessary handle
error scenarios appropriately.