Improve hit detection code, use complex boundry to create cutout for menu view
Needed to determine if an event should be directed to the menu layer. Created a
complex boundry, which is composed of normal boundries, to help determine if an
event should go there. We do hit detection against the complex boundry. The lead
to moving the hit method to boundry, which helped clean up some of the hit
detection code for the canvas, and led to deleting the hittable mixin. Trying to
use more javascript, using the apply method to bleed arguments through. Creates
an extra variable reference, but the addtional fexablitiy is useful.