Map event to be relative to menu 0,0
Establish grid corrdinate system 0,0 point to be left offset according the when
the container is centered to. Need to handel attach and window resize events to
key the retrieved value in sync with the actual value. The attach event is new,
and give a point to measure on attached into the dom, which is the first point
we can measure for a value. There is probably a cleaner way of doing this.
Either injecting the el, or having a singular event that happens after all the
views have been constructred. Attach works alright, but ran into some naming
conflicts. We'll see if it evolves.