Z-index fixes, and hacks to maintain order between grouping and ungrouping
I figure the z-index is getting lost in grouping and ungrouping due to the order
in which things are pulled out and added back. By reversing the order in two
places, I am able to avoid the issue of losing order in the test cases I ran,
but this is symptomatic of needing to revisit and improve the approach to
maintaining z-index. We just need better support and and more straight forward
and robust solution.