Raw implementation of photo captions
A new attribute is added to photo content items, namely the caption_height
attribute. This is used for computing the proper height of the photo, which is
now inside of a div container and the various heights are calculated and set
during resize to implement the expected behavior. The user can now toggle on and
off the caption, updated and still has resize and move functionalty on the
photo.