Basic video display on canvas
These changes allow the video to be displayed on the canvas, along with the play
button overlay and the cover image.

I had to simplify the expectation, that we will simply be putting the raw html
from the server side render of the "Video feature". Previously, I was using an
angular directive to do this, which was cool and handy, but is unfortuanly
conflicting with the composer at the moment, and getting these two to play nice,
will have to be something scheduled later on down the road.

For the most part, it works well. It is recommended that only mp4 videos are
used, as I did not find success with Quicktime videos. Another video format may
work well too. This is an area for further research.

Currently, the default video widget doesn't nessecary allow to to scan trough
the video and jump around. It can also crash. More details, and perhaps some
soltuion might be found here. http://stackoverflow.com/questions/28008564/streaming-mp4-in-chrome-with-rails-nginx-and-send-file#

I added some of this code, and was finding success with it, but ultamatly have
let it be commented out for now, to err on the side of simplicity. I figure we
can tweek in subsequent commits.

It will be important to have video working very well, with all of the features
user have come to expect.

This may require the development of our own video frame, but more realistically,
I think we just need to get the formula right on serving the video. These issues
are possibly linked to using send file, and streaming the video, and so
hopefully, and quite possibly, they will disappear when using the compiled
version of the publication.