Use set aspect ratio when drag bottom right handle of video
Sort of a messy commit, but proof of concept, using the video aspect ratio when
resizing video.

Should always honer aspect ratio, no matter what handle is be used.

Also, looking past the fact that the video belongs_to isn't being properly set,
so simply looking up by video name to render the aspect ratio. This works for
the time being, but should be corrected in the future.

Cheers and love.