First shot at adding a video caption
A little different than photo captions, but roughly the same.

Currently the captions render below the video, outside of the frame. This is
different than photos where they render within the frame. This may or may not
change in the future.

Works for the most part, and helps to maintain the aspect ratio of the video,
but violates the idea that all of the content should appear within the boundry.

Will be a little more complex to honer the boundry and keep the caption in the
frame. This work for now, at least, so that we can have captions on video.

Also a side note that the positioning of the element is a little less than
standard, but just working with what we got and trying to get satisfactory
results.

Cheers!