Fix videos being included in the compiled print
The videos were being skipped in the print because, for the time being, the are
being related by name, rather then by a proper if relationship. This was done
for convience in the prototype, but the code should be updated and corrected in
te future.

By following the name look up convention, the videos and cover image are
porperly included. Hallelujah!

Love <3