Better assign page, add some docs
Now when dragging an item across the page boundry, it is reassigned, and a
change made to the structure, namly: which page contains the item in it's canvas
items array. This is done with the add and remove canvas item methods on the
page view. Also, a renderPanel is sent to the pages panel to ensure that
reflects the changes. Lastly, we are ensuring to update the assigned page in the
data.

Add some documentation to the doc directory. This is the beginning of a
documentation effort to fully documentation the application, both in design and
function, and also to help people get up and running with the tool, for example,
installation instructions.