Much improved detection in panel hovering
Before we were missing mouseout events on the panels, which caused the the
overlaying capture view layer to to not be reengaged, and this disable the
functionality of the UI.

In this change, we continually check at the panel view layer, when mouse events
are received there, (meaning they haven't been eaten by the panel), if they
should be considered still over a panel (sometimes they can slip through). If a
mouse event is not over a panel, then the panel view layer reengages the capture
view layer and unhovers it's children. This is good, and works well.

There are still contingencies in place, such as clicking on the panel view layer
to reset the composer, but those should now be not needed. Yay!