Remove auth in many respects, begin working on a blank action
Going to move towards document based auth, for multiuser hosting. The
application itself is less critical to secure, and can work in an annonmous
mode. The initial load should be a blank document, and the user can go from
there. To sign in with an account will allow the user to own documents, and
these will not be shared so freely.