Fix placement of marquee on selected grouped item
When items are grouped, and subsequently selected, the offset of the group needs
to be considered when drawing the marquee. Previously, it was impossible to
select something within a group, but now that this is possible using the pages
panel, it is important that the offset of the group, or groups, which the item
is nested within, is taken into account.